pyproject.toml
[project]
name = "pydiction"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "quantike", email = "ike.chasse@gmail.com" }
]
dependencies = [
    "websockets>=13.1",
    "asyncio>=3.4.3",
    "requests>=2.32.3",
    "cryptography>=43.0.1",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "jupyter>=1.1.1",
    "polars>=1.8.2",
    "matplotlib>=3.9.2",
    "connectorx>=0.3.3",
    "pyarrow>=17.0.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pydiction"]
